bug: fix list/find-fix useeffect, add list/algfind
- Fixed issue in useEffect.
- Refactored to move useEffect to operations
- Modified list/find to route to FindRequestsContainer.tsg instead of
  FindRequestsAlgoliaContainer
- Created new route list/algfind to route to
  FindRequestsAlgoliaContainer
