+++ web-client/src/modules/myRequests/constants.tsx
@@ -8,6 +8,8 @@ export enum PostTabTypes {
   offers = 'offers',
 }
 
+// TODO: (es) Pick between these two methods
+export const AlgFindRequestsLocation = new Location(`${RootUrl}/algfind`);
 export const FindRequestsLocation = new Location(`${RootUrl}/find`);
 export const MyPostsLocation = new Location(`${RootUrl}/my/:postType`, {
   postType: Yup.string().required(),
