+++ web-client/src/modules/myRequests/routes.tsx
@@ -2,10 +2,19 @@ import React, { lazy, ReactElement, Suspense } from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
 
 import LoadingWrapper from '../../components/LoadingComponent/LoadingComponent';
-import { FindRequestsLocation, MyPostsLocation } from './constants';
+import {
+  AlgFindRequestsLocation,
+  FindRequestsLocation,
+  MyPostsLocation,
+} from './constants';
+
+// TODO: (es) Choose between Alg and NonAlg container
+const AlgFindRequestsContainer = lazy(() =>
+  import('./containers/FindRequestsAlgoliaContainer'),
+);
 
 const FindRequestsContainer = lazy(() =>
-  import('./containers/FindRequestsAlgoliaContainer'),
+  import('./containers/FindRequestsContainer'),
 );
 
 const MyPostsTabs = lazy(() =>
@@ -15,9 +24,15 @@ const MyPostsTabs = lazy(() =>
 const Routes = (): ReactElement => (
   <Suspense fallback={<LoadingWrapper />}>
     <Switch>
+      {/* TODO: (es) Decide on algolia or non algolia */}
       <Route
         path={FindRequestsLocation.path}
         component={FindRequestsContainer}
+        exact
+      />
+      <Route
+        path={AlgFindRequestsLocation.path}
+        component={AlgFindRequestsContainer}
         exact
       />
       <Route path={MyPostsLocation.path} component={MyPostsTabs} />
